禁止老师帮学生提交申请

+++ tm-wx-web/grails-app/domain/cn/edu/bnuz/bell/wx/eto/ScoreEto.groovy
@@ -10,7 +10,7 @@ class ScoreEto {
     String courseName
     String credit
     String score
-    Integer scoreNumber
+    BigDecimal scoreNumber
 
     static mapping = {
         table       name: 'et_score', schema: 'tm_wx'
